+++ NEWS.md
@@ -1,6 +1,7 @@
 # EDBdata (development version)
 
 * Added `phyto_edb` - phytoplankton community data collected by DWR-EMP from 2014-2021 for the stations within the designated EDB regions
+* Renamed the `sat_ci_count_fr_mil` data set to `hab_sat_fr_mil` to be more descriptive
 
 # EDBdata 0.1.0